complete sub preferences fragment version
