+++ app/src/main/res/xml/settings_general.xml
@@ -23,13 +23,34 @@
 
     </PreferenceCategory>
 
-    <PreferenceCategory app:title="Device Settings">
+    <PreferenceCategory >
         <SwitchPreference
             app:defaultValue="false"
             app:icon="@android:drawable/ic_lock_power_off"
             app:key="@string/key_device_status"
             app:summary="Disconnected"
             app:title="Device Status:" />
+    </PreferenceCategory>
+    <PreferenceCategory
+        android:dependency="@string/key_device_status"
+        app:title="Device Settings">
+
+        <EditTextPreference
+            android:inputType="number"
+            app:defaultValue="5"
+            app:icon="@android:drawable/ic_menu_save"
+            app:key="@string/key_setting_stime"
+            app:summary="@string/summary_stime"
+            app:title="Sample Time Interval" />
+
+        <ListPreference
+            android:defaultValue="0"
+            android:entries="@array/pref_dtype_entries"
+            android:entryValues="@array/pref_dtype_values"
+            android:key="@string/key_setting_dtype"
+            android:title="Sensor type"
+            app:useSimpleSummaryProvider="true" />
+
 
         <Preference
             android:dependency="@string/key_device_status"
@@ -44,7 +65,6 @@
             app:summary="Some device actions and settings"
             app:title="CanAirIO device tools">
         </Preference>
-
     </PreferenceCategory>
 
 </PreferenceScreen>
